Move ConfigurableParam
Move the configurable params to a more fundamental
library so that we can use it better in other modules.

Move Rubens NameConf to DataFormats/Detector/Common
