+++ Common/Utils/CMakeLists.txt
@@ -12,6 +12,7 @@ o2_add_library(CommonUtils
                SOURCES src/TreeStream.cxx src/TreeStreamRedirector.cxx
                        src/RootChain.cxx src/CompStream.cxx src/ShmManager.cxx
 	               src/ValueMonitor.cxx
+                       src/ConfigurableParamHelper.cxx src/ConfigurableParam.cxx
                PUBLIC_LINK_LIBRARIES ROOT::Hist ROOT::Tree Boost::iostreams O2::CommonDataFormat O2::Headers
                                      FairLogger::FairLogger)
 
@@ -24,7 +25,9 @@ o2_target_root_dictionary(CommonUtils
                                   include/CommonUtils/RngHelper.h
                                   include/CommonUtils/StringUtils.h
                                   include/CommonUtils/ValueMonitor.h
-				  include/CommonUtils/MemFileHelper.h)
+				  include/CommonUtils/MemFileHelper.h
+                                  include/CommonUtils/ConfigurableParam.h
+                                  include/CommonUtils/ConfigurableParamHelper.h)
 
 o2_add_test(TreeStream
             COMPONENT_NAME CommonUtils
