+++ Common/Utils/include/CommonUtils/ConfigurableParamHelper.h
@@ -13,7 +13,7 @@
 #ifndef COMMON_SIMCONFIG_INCLUDE_SIMCONFIG_CONFIGURABLEPARAMHELPER_H_
 #define COMMON_SIMCONFIG_INCLUDE_SIMCONFIG_CONFIGURABLEPARAMHELPER_H_
 
-#include "SimConfig/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParam.h"
 #include "TClass.h"
 #include <type_traits>
 #include <typeinfo>
