+++ DataFormats/Detectors/Common/CMakeLists.txt
@@ -10,11 +10,13 @@
 
 o2_add_library(DetectorsCommonDataFormats
                SOURCES src/DetID.cxx src/AlignParam.cxx src/DetMatrixCache.cxx
+                       src/NameConf.cxx
                PUBLIC_LINK_LIBRARIES
                ROOT::Core
                ROOT::Geom
                O2::GPUCommon
-               O2::MathUtils)
+               O2::MathUtils
+               O2::CommonUtils)
 
 o2_target_root_dictionary(
   DetectorsCommonDataFormats
@@ -22,6 +24,7 @@ o2_target_root_dictionary(
           include/DetectorsCommonDataFormats/FileNameGenerator.h
           include/DetectorsCommonDataFormats/SimTraits.h
           include/DetectorsCommonDataFormats/AlignParam.h
+          include/DetectorsCommonDataFormats/NameConf.h
           include/DetectorsCommonDataFormats/DetMatrixCache.h)
 
 o2_add_test(DetID
