+++ DataFormats/Detectors/Common/src/DetectorsCommonDataFormatsLinkDef.h
@@ -24,4 +24,7 @@
 #pragma link C++ class o2::filenames::SimFileNameGenerator;
 #pragma link C++ class o2::detectors::SimTraits + ;
 
+#pragma link C++ class o2::base::NameConf + ;
+#pragma link C++ class o2::conf::ConfigurableParamHelper < o2::base::NameConf> + ;
+
 #endif
