+++ Detectors/Base/CMakeLists.txt
@@ -17,7 +17,6 @@ o2_add_library(DetectorsBase
                        src/MatLayerCylSet.cxx
                        src/Ray.cxx
 		       src/DCAFitter.cxx
-		       src/NameConf.cxx
                PUBLIC_LINK_LIBRARIES FairRoot::Base
                                      O2::CommonUtils
                                      O2::DetectorsCommonDataFormats
@@ -38,8 +37,7 @@ o2_target_root_dictionary(DetectorsBase
                                   include/DetectorsBase/Ray.h
                                   include/DetectorsBase/MatCell.h
                                   include/DetectorsBase/MatLayerCyl.h
-                                  include/DetectorsBase/MatLayerCylSet.h
-				  include/DetectorsBase/NameConf.h)
+                                  include/DetectorsBase/MatLayerCylSet.h)
 
 if(BUILD_SIMULATION)
   o2_add_test(
