+++ Detectors/Base/src/DetectorsBaseLinkDef.h
@@ -29,7 +29,4 @@
 
 #pragma link C++ class o2::base::DCAFitter + ;
 
-#pragma link C++ class o2::base::NameConf + ;
-#pragma link C++ class o2::conf::ConfigurableParamHelper < o2::base::NameConf> + ;
-
 #endif
