+++ Detectors/FIT/FV0/simulation/include/FV0Simulation/FV0DigParam.h
@@ -11,8 +11,8 @@
 #ifndef O2_FV0_SIMPARAMS_H_
 #define O2_FV0_SIMPARAMS_H_
 
-#include "SimConfig/ConfigurableParam.h"
-#include "SimConfig/ConfigurableParamHelper.h"
+#include "CommonUtils/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParamHelper.h"
 
 namespace o2
 {
