+++ Detectors/FIT/workflow/src/fit-reco-workflow.cxx
@@ -9,7 +9,7 @@
 // or submit itself to any jurisdiction.
 
 #include "FITWorkflow/RecoWorkflow.h"
-#include "SimConfig/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParam.h"
 
 using namespace o2::framework;
