+++ Detectors/MUON/MID/Workflow/src/mid-digits-to-raw.cxx
@@ -16,7 +16,7 @@
 #include <string>
 #include <vector>
 #include "Framework/Variant.h"
-#include "SimConfig/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParam.h"
 #include "MIDWorkflow/DigitsToRawWorkflow.h"
 
 using namespace o2::framework;
