+++ Detectors/TPC/base/include/TPCBase/ParameterDetector.h
@@ -17,8 +17,8 @@
 
 #include <array>
 #include "DataFormatsTPC/Defs.h"
-#include "SimConfig/ConfigurableParam.h"
-#include "SimConfig/ConfigurableParamHelper.h"
+#include "CommonUtils/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParamHelper.h"
 
 namespace o2
 {
