+++ Detectors/TPC/base/include/TPCBase/ParameterGEM.h
@@ -21,8 +21,8 @@
 
 #include <array>
 #include <cmath>
-#include "SimConfig/ConfigurableParam.h"
-#include "SimConfig/ConfigurableParamHelper.h"
+#include "CommonUtils/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParamHelper.h"
 
 namespace o2
 {
