+++ Detectors/TRD/simulation/include/TRDSimulation/TRDSimParams.h
@@ -13,8 +13,8 @@
 
 // Global parameters for TRD simulation / digitization
 
-#include "SimConfig/ConfigurableParam.h"
-#include "SimConfig/ConfigurableParamHelper.h"
+#include "CommonUtils/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParamHelper.h"
 
 namespace o2
 {
