+++ Generators/include/Generators/InteractionDiamondParam.h
@@ -13,8 +13,8 @@
 #ifndef ALICEO2_EVENTGEN_INTERACTIONDIAMONDPARAM_H_
 #define ALICEO2_EVENTGEN_INTERACTIONDIAMONDPARAM_H_
 
-#include "SimConfig/ConfigurableParam.h"
-#include "SimConfig/ConfigurableParamHelper.h"
+#include "CommonUtils/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParamHelper.h"
 
 namespace o2
 {
