+++ Generators/include/Generators/TriggerParticleParam.h
@@ -13,8 +13,8 @@
 #ifndef ALICEO2_EVENTGEN_TRIGGERPARTICLEPARAM_H_
 #define ALICEO2_EVENTGEN_TRIGGERPARTICLEPARAM_H_
 
-#include "SimConfig/ConfigurableParam.h"
-#include "SimConfig/ConfigurableParamHelper.h"
+#include "CommonUtils/ConfigurableParam.h"
+#include "CommonUtils/ConfigurableParamHelper.h"
 
 namespace o2
 {
